Gossip integration

+++ subnet/cli/run_node.py
@@ -2,22 +2,33 @@
 
 import argparse
 import logging
-import sys
 import os
 from pathlib import Path
+import random
+import secrets
+import sys
+import time
 
+from dotenv import load_dotenv
+from libp2p.crypto.ed25519 import (
+    Ed25519PrivateKey,
+    create_new_key_pair,
+)
+from libp2p.crypto.keys import KeyPair
+from libp2p.crypto.secp256k1 import Secp256k1PrivateKey
+from libp2p.peer.id import ID as PeerID
+from libp2p.peer.pb import crypto_pb2
+from substrateinterface import (
+    Keypair as SubstrateKeypair,
+    KeypairType,
+)
 import trio
 
-from subnet.server.server_v2 import Server
-import random
-from subnet.hypertensor.mock.local_chain_functions import LocalMockHypertensor
+from subnet.db.database import RocksDB
 from subnet.hypertensor.chain_functions import Hypertensor, KeypairFrom
-from libp2p.crypto.keys import KeyPair
-from libp2p.peer.pb import crypto_pb2
-from libp2p.crypto.ed25519 import Ed25519PrivateKey
-from libp2p.peer.id import ID as PeerID
-from substrateinterface import Keypair as SubstrateKeypair, KeypairType
-from dotenv import load_dotenv
+from subnet.hypertensor.mock.local_chain_functions import LocalMockHypertensor
+from subnet.server.server import Server
+from subnet.utils.crypto.store_key import get_key_pair
 
 load_dotenv(os.path.join(Path.cwd(), ".env"))
 
@@ -39,21 +50,128 @@ def parse_args() -> argparse.Namespace:
         formatter_class=argparse.RawDescriptionHelpFormatter,
         epilog="""
 Examples:
-  # Run a standalone node
-  python -m subnet.cli.run_node
+# Run locally with no RPC connection
 
-  # Run a node connecting to bootstrap peers
-  python -m subnet.cli.run_node --bootstrap /ip4/127.0.0.1/tcp/31330/p2p/QmBootstrapPeerID
+# Start bootnode (or start bootnode through `run_bootnode`)
 
-  # Connect to multiple bootstrap peers
-  python -m subnet.cli.run_node \\
-    --bootstrap /ip4/192.168.1.100/tcp/31330/p2p/QmPeer1 \\
-    --bootstrap /ip4/192.168.1.101/tcp/31330/p2p/QmPeer2
+python -m subnet.cli.run_node \
+--private_key_path bootnode.key \
+--port 38960 \
+--subnet_id 1 \
+--subnet_node_id 1 \
+--no_blockchain_rpc \
+--is_bootstrap
 
-  # Run a node with an identity file
-  python -m subnet.cli.run_node --identity_path alith-ed25519.key --port 38960 --bootstrap /ip4/127.0.0.1/tcp/38959/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF --subnet_id 1 --subnet_node_id 1 --no_blockchain_rpc
-  python -m subnet.cli.run_node --identity_path baltathar-ed25519.key --port 38961 --bootstrap /ip4/127.0.0.1/tcp/38959/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF --subnet_id 1 --subnet_node_id 2 --no_blockchain_rpc
-  python -m subnet.cli.run_node --identity_path charleth-ed25519.key --port 38962 --bootstrap /ip4/127.0.0.1/tcp/38959/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF --subnet_id 1 --subnet_node_id 3 --no_blockchain_rpc
+# Connect to bootnode
+
+python -m subnet.cli.run_node \
+--private_key_path alith.key \
+--port 38961 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 1 \
+--no_blockchain_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path baltathar.key \
+--port 38962 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 2 \
+--no_blockchain_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path charleth.key \
+--port 38963 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 3 \
+--no_blockchain_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path dorothy.key \
+--port 38964 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 4 \
+--no_blockchain_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path faith.key \
+--port 38965 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 5 \
+--no_blockchain_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path george.key \
+--port 38966 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 6 \
+--no_blockchain_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path harry.key \
+--port 38967 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 7 \
+--no_blockchain_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path ian.key \
+--port 38968 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 8 \
+--no_blockchain_rpc
+
+
+# Run locally with local RPC connection
+
+# Start bootnode (or start bootnode through `run_bootnode`)
+
+- Register subnet
+- Register subnet nodes
+
+# Start nodes
+
+python -m subnet.cli.run_node \
+--private_key_path bootnode.key \
+--port 38960 \
+--subnet_id 1 \
+--subnet_node_id 1 \
+--local_rpc \
+--is_bootstrap
+
+# Connect to bootnode
+
+python -m subnet.cli.run_node \
+--private_key_path alith.key \
+--port 38961 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 1 \
+--local_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path baltathar.key \
+--port 38962 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 2 \
+--local_rpc
+
+python -m subnet.cli.run_node \
+--private_key_path charleth.key \
+--port 38963 \
+--bootstrap /ip4/127.0.0.1/tcp/38960/p2p/12D3KooWLGmub3LXuKQixBD5XwNW4PtSfnrysYzqs1oj19HxMUCF \
+--subnet_id 1 \
+--subnet_node_id 3 \
+--local_rpc
+
         """,
     )
 
@@ -80,23 +198,30 @@ Examples:
     )
 
     parser.add_argument(
-        "--identity_path", type=str, default=None, help="Path to the identity file. "
+        "--is_bootstrap",
+        action="store_true",
+        help="Start a bootnode that doesn't publish messages or run consensus. ",
     )
 
+    parser.add_argument("--base_path", type=str, default=None, help="Specify custom base path")
+
     parser.add_argument(
-        "--subnet_id", type=int, default=1, help="Subnet ID this node belongs to. "
+        "--private_key_path",
+        type=str,
+        default=None,
+        help="Path to the private key file. ",
     )
+
+    parser.add_argument("--subnet_id", type=int, default=1, help="Subnet ID this node belongs to. ")
 
     parser.add_argument(
         "--subnet_node_id",
         type=int,
-        default=1,
+        default=0,
         help="Subnet node ID this node belongs to. ",
     )
 
-    parser.add_argument(
-        "--no_blockchain_rpc", action="store_true", help="[Testing] Run with no RPC"
-    )
+    parser.add_argument("--no_blockchain_rpc", action="store_true", help="[Testing] Run with no RPC")
 
     parser.add_argument(
         "--local_rpc",
@@ -151,15 +276,21 @@ def main() -> None:
     else:
         logger.info("Running as standalone node (no bootstrap peers)")
 
-    with open(f"{args.identity_path}", "rb") as f:
-        data = f.read()
-    private_key = crypto_pb2.PrivateKey.FromString(data)
-    ed25519_private_key = Ed25519PrivateKey.from_bytes(private_key.Data)
-    public_key = ed25519_private_key.get_public_key()
-    key_pair = KeyPair(ed25519_private_key, public_key)
+    if args.private_key_path is None:
+        key_pair = create_new_key_pair(secrets.token_bytes(32))
+    else:
+        key_pair = get_key_pair(args.private_key_path)
+
+    if not args.base_path:
+        base_path = f"/tmp/{random.randint(100, 1000000)}"
+    else:
+        base_path = args.base_path
+
+    db = RocksDB(base_path, args.subnet_id)
 
     hotkey = None
     start_epoch = None
+
     if not args.no_blockchain_rpc:
         if args.local_rpc:
             rpc = os.getenv("LOCAL_RPC")
@@ -168,15 +299,11 @@ def main() -> None:
 
         if args.phrase is not None:
             hypertensor = Hypertensor(rpc, args.phrase)
-            substrate_keypair = SubstrateKeypair.create_from_mnemonic(
-                args.phrase, crypto_type=KeypairType.ECDSA
-            )
+            substrate_keypair = SubstrateKeypair.create_from_mnemonic(args.phrase, crypto_type=KeypairType.ECDSA)
             hotkey = substrate_keypair.ss58_address
             logger.info(f"hotkey: {hotkey}")
         elif args.tensor_private_key is not None:
-            hypertensor = Hypertensor(
-                rpc, args.tensor_private_key, KeypairFrom.PRIVATE_KEY
-            )
+            hypertensor = Hypertensor(rpc, args.tensor_private_key, KeypairFrom.PRIVATE_KEY)
             substrate_keypair = SubstrateKeypair.create_from_private_key(
                 args.tensor_private_key, crypto_type=KeypairType.ECDSA
             )
@@ -186,44 +313,98 @@ def main() -> None:
             # Default to using PHRASE if no other options are provided
             hypertensor = Hypertensor(rpc, PHRASE)
 
-        if hotkey is not None:
-            result = hypertensor.interface.query("System", "Account", [hotkey])
-            balance = result.value["data"]["free"]
-            assert balance >= 500, (
-                f"Hotkey must have at least 0.0000000000000005 TENSOR to be a live account, balance is {float(balance / 1e18)}"
-            )
+        if not args.is_bootstrap:
+            if hotkey is not None:
+                result = hypertensor.interface.query("System", "Account", [hotkey])
+                balance = result.value["data"]["free"]
+                assert balance >= 500, (
+                    f"Hotkey must have at least 0.0000000000000005 TENSOR to be a live account, balance is {float(balance / 1e18)}"  # noqa: E501
+                )
 
-        # Check subnet node exists
-        subnet_node_info = hypertensor.get_formatted_get_subnet_node_info(
-            args.subnet_id, args.subnet_node_id
-        )
-        if subnet_node_info is None:
-            raise Exception("Subnet node does not exist")
+            # Check subnet node exists
+            subnet_node_info = hypertensor.get_formatted_get_subnet_node_info(args.subnet_id, args.subnet_node_id)
+            if subnet_node_info is None:
+                raise Exception("Subnet node does not exist")
 
-        start_epoch = subnet_node_info.classification["start_epoch"]
-        if start_epoch is None:
-            raise Exception("Subnet node start epoch is None")
+            if subnet_node_info.hotkey.lower() != hotkey.lower():
+                raise Exception(
+                    f"Subnet node hotkey does not match. Expected: {subnet_node_info.hotkey}, Actual: {hotkey}"
+                )
+
+            if PeerID.from_pubkey(key_pair.public_key).__eq__(subnet_node_info.peer_id):
+                logger.warning(
+                    "Subnet node peer ID does not match. This can be ignored if running a bootnode or client peer. "
+                    f"Expected: {subnet_node_info.peer_id}, Actual: {PeerID.from_pubkey(key_pair.public_key).to_base58()}"  # noqa: E501
+                )
+
+            if PeerID.from_pubkey(key_pair.public_key).__eq__(subnet_node_info.bootnode_peer_id):
+                logger.warning(
+                    "Subnet node bootnode peer ID does not match. This can be ignored if you're not running a bootnode. "  # noqa: E501
+                    f"Expected: {subnet_node_info.bootnode_peer_id}, Actual: {PeerID.from_pubkey(key_pair.public_key).to_base58()}"  # noqa: E501
+                )
+
+            if PeerID.from_pubkey(key_pair.public_key).__eq__(subnet_node_info.client_peer_id):
+                logger.warning(
+                    "Subnet node client peer ID does not match. This can be ignored if you're not running a client peer. "  # noqa: E501
+                    f"Expected: {subnet_node_info.client_peer_id}, Actual: {PeerID.from_pubkey(key_pair.public_key).to_base58()}"  # noqa: E501
+                )
+
+            start_epoch = subnet_node_info.classification["start_epoch"]
+            if start_epoch is None:
+                raise Exception("Subnet node start epoch is None")
     else:
         # Run mock hypertensor blockchain for testing
+        # This is a shared database between all local nodes
         hypertensor = LocalMockHypertensor(
             subnet_id=args.subnet_id,
             peer_id=PeerID.from_pubkey(key_pair.public_key),
-            subnet_node_id=args.subnet_node_id,
+            subnet_node_id=args.subnet_node_id if not args.is_bootstrap else 0,
             coldkey="",
             hotkey="",
             bootnode_peer_id="",
             client_peer_id="",
             reset_db=True if not args.bootstrap else False,
+            insert_mock_overwatch_node=True if not args.bootstrap else False,
         )
+
+    # Wait to start the node until the node is fully registered on-chain
+    # NOTE: Once a node registers on-chain, it will not be considered fully registered to other nodes
+    # until the following epoch to ensure it starts on a fresh epoch.
+    if start_epoch is not None:
+        slot = hypertensor.get_subnet_slot(args.subnet_id)
+        slot = int(str(slot))
+        subnet_epoch_data = hypertensor.get_subnet_epoch_data(slot)
+        current_epoch = subnet_epoch_data.epoch
+        logger.info(f"Current epoch is {current_epoch}")
+        if current_epoch < start_epoch:
+            logger.info(
+                "Keep this running and the node will automatically join the subnet once it's fully registered on-chain"
+            )
+            logger.info(f"Subnet node start epoch is {start_epoch}")
+            while current_epoch < start_epoch:
+                subnet_epoch_data = hypertensor.get_subnet_epoch_data(slot)
+                current_epoch = subnet_epoch_data.epoch
+                logger.info(f"Current epoch is {current_epoch}")
+                if current_epoch >= start_epoch:
+                    break
+
+                seconds_remaining = subnet_epoch_data.seconds_remaining
+                logger.info(
+                    f"Checking next epoch to see if we can join the subnet, sleeping for {seconds_remaining} seconds"
+                )
+                time.sleep(seconds_remaining)
+        logger.info("Subnet node is about to join the subnets DHT")
 
     try:
         server = Server(
             port=port,
             bootstrap_addrs=args.bootstrap,
             key_pair=key_pair,
+            db=db,
             subnet_id=args.subnet_id,
             subnet_node_id=args.subnet_node_id,
             hypertensor=hypertensor,
+            is_bootstrap=args.is_bootstrap,
         )
         trio.run(server.run)
     except KeyboardInterrupt:
@@ -233,12 +414,13 @@ def main() -> None:
         logger.error(f"Fatal error: {e}", exc_info=True)
         sys.exit(1)
     finally:
-        try:
-            # Delete subnet node from mock db if it was created
-            hypertensor.db.delete_subnet_node(args.subnet_id, args.subnet_node_id)
-            logger.info("Removed subnet node from mock db")
-        except Exception as e:
-            logger.error(f"Failed to delete subnet node from mock db: {e}")
+        if isinstance(hypertensor, LocalMockHypertensor):
+            try:
+                # Delete subnet node from mock db if it was created
+                hypertensor.db.delete_subnet_node(args.subnet_id, args.subnet_node_id)
+                logger.info("Removed subnet node from mock db")
+            except Exception as e:
+                logger.error(f"Failed to delete subnet node from mock db: {e}")
 
 
 if __name__ == "__main__":
